+++ Source/Experimental/ShadowViewport/OSGPCFShadowMap.cpp
@@ -3965,7 +3965,7 @@ void PCFShadowMap::createShadowFactorMap(RenderActionBase *action)
         {
             UInt32  renderTimes = 1;
             if(lightCounter > 4)
-                renderTimes = ceil(Real32(lightCounter) / 4.0f);
+                renderTimes = UInt32(ceil(Real32(lightCounter) / 4.0f));
 
             for(UInt32 i = 0;i < renderTimes;i++)
             {
@@ -4606,7 +4606,7 @@ void PCFShadowMap::createShadowFactorMapFBO(RenderActionBase *action)
     {
         UInt32  renderTimes = 1;
         if(lightCounter > 4)
-            renderTimes = ceil(Real32(lightCounter) / 4.0f);
+            renderTimes = UInt32(ceil(Real32(lightCounter) / 4.0f));
 
         for(UInt32 i = 0;i < renderTimes;i++)
         {
