+++ Source/Experimental/ShadowViewport/OSGPerspectiveShadowMap.cpp
@@ -3041,7 +3041,7 @@ void PerspectiveShadowMap::createShadowFactorMap(RenderActionBase *action)
     {
         UInt32  renderTimes = 1;
         if(lightCounter > 7)
-            renderTimes = ceil(Real32(lightCounter) / 7.0f);
+            renderTimes = UInt32(ceil(Real32(lightCounter) / 7.0f));
 
         for(UInt32 i = 0;i < renderTimes;i++)
         {
@@ -3726,7 +3726,7 @@ void PerspectiveShadowMap::createShadowFactorMapFBO(RenderActionBase *action)
 
         UInt32  renderTimes = 1;
         if(lightCounter > 7)
-            renderTimes = ceil(Real32(lightCounter) / 7.0f);
+            renderTimes = UInt32(ceil(Real32(lightCounter) / 7.0f));
 
         for(UInt32 i = 0;i < renderTimes;i++)
         {
